Move and rename files.
Files have been moved from opm/core/grid to opm/grid.
Renamed grid.h -> UnstructuredGrid.h.
