+++ opm/core/utility/VelocityInterpolation.cpp
@@ -19,7 +19,7 @@
 
 #include "config.h"
 #include <opm/core/utility/VelocityInterpolation.hpp>
-#include <opm/core/grid.h>
+#include <opm/grid/UnstructuredGrid.h>
 #include <opm/common/utility/numeric/blas_lapack.h>
 
 #include <iostream>
