+++ opm/grid/CellQuadrature.hpp
@@ -20,7 +20,7 @@
 #ifndef OPM_CELLQUADRATURE_HEADER_INCLUDED
 #define OPM_CELLQUADRATURE_HEADER_INCLUDED
 
-#include <opm/core/grid.h>
+#include <opm/grid/UnstructuredGrid.h>
 #include <opm/grid/utility/ErrorMacros.hpp>
 #include <algorithm>
 #include <cmath>
